Ongoing refactoring of effect system
Due to "small" mistake effects didn't really worked when there were multiple targets or multiple effects of the same type, as the effect and handle were reused twice.

Now effects of the same type are added to Queue and removed from oldest to newest. By default one effect is removed but it is possible to specific amount.

It is not yet finished and there is still ongoing cleanup for unsed params/functions.